+++ Library/PAX_MAHOROBA/Core/SimulationController.hpp
@@ -120,7 +120,7 @@ private:
         if (total_steps_ <= 0) {
             return false;
         }
-        return koyomi.steps.getDay() >= static_cast<std::size_t>(total_steps_);
+        return koyomi.steps.getDay() >= static_cast<std::uint_least32_t>(total_steps_);
     }
 
     /// @brief 次の実行サイクルを開始
